fix: make FillMesh instanced to enhance perf #1359
